Extract internal types from `types` module
